+++ src/ambition/version_details.cc
@@ -27,6 +27,7 @@ std::vector<VersionUpdates> versionDetails = {
         .items = {
           _("Show Spy combat (in Forts) or primary skill (in other Firms) level in the worker list by moving the spy icon to the portrait."),
           _("Added thousands seperators to treasury and food store amounts, to make reading them easier.  (Currently, this uses groups of three with a comma as the seperator for everyone.  Proper localisation will be added later.)"),
+          _("Show the war machine technology level on the portrait, instead of as part of the name, on the War Factory build list.  This frees up some room for more information."),
         },
       },
     },